+++ calkit/cli/sync.py
@@ -5,7 +5,9 @@ from __future__ import annotations
 from typing import Callable
 
 import typer
+from typing_extensions import Annotated
 
+import calkit
 from calkit.cli import AliasGroup
 
 sync_app = typer.Typer(cls=AliasGroup, no_args_is_help=True)
@@ -51,3 +53,53 @@ def sync_all() -> None:
     # Exit with an error if any target failed so callers can react.
     if failures:
         raise typer.Exit(1)
+
+
+def _is_git_configured() -> bool:
+    """Check whether the Git repository is configured for syncing."""
+    try:
+        repo = calkit.git.get_repo()
+        return len(repo.remotes) > 0
+    except Exception:
+        return False
+
+
+@sync_app.command(name="git")
+def sync_git(
+    no_check_auth: Annotated[bool, typer.Option("--no-check-auth")] = False,
+) -> None:
+    """Sync the Git repository by pulling and then pushing."""
+    from calkit.cli.main.core import pull, push
+
+    pull(no_dvc=True, no_check_auth=no_check_auth)
+    push(no_dvc=True, no_check_auth=no_check_auth)
+
+
+register_sync_target("git", sync_git, _is_git_configured)
+
+
+def _is_dvc_configured() -> bool:
+    """Check whether DVC is configured for syncing.
+
+    A DVC repo alone is not enough to sync: pulling or pushing requires at
+    least one remote. We therefore check both that a DVC repo exists and that
+    it has remotes configured.
+    """
+    try:
+        return len(calkit.dvc.get_remotes()) > 0
+    except Exception:
+        return False
+
+
+@sync_app.command(name="dvc")
+def sync_dvc(
+    no_check_auth: Annotated[bool, typer.Option("--no-check-auth")] = False,
+) -> None:
+    """Sync the DVC repository by pulling and then pushing."""
+    from calkit.cli.main.core import pull, push
+
+    pull(no_git=True, no_check_auth=no_check_auth)
+    push(no_git=True, no_check_auth=no_check_auth)
+
+
+register_sync_target("dvc", sync_dvc, _is_dvc_configured)
